feat(parse): can now import the portfolio structure
