+++ finalynx/copilot/copilot.py
@@ -3,7 +3,7 @@
 TODO Dummy class for now, check back later or help us by contributing!
 ```
 """
-from ..portfolio.portfolio import Portfolio
+from ..portfolio.folder import Portfolio
 
 
 class Copilot:
